Add recipe-book service methods
Injected the recipe-book service into the recipe-book component and it's children.
Replaced the input/outpt functionality of selecting a recipe with a leaner service based way of doing it.
This uses a service to assigned the selected recipe to an subscribable property in the service.
This property can be accessed by any of the recipe components.